data_qa_new/requirements.txt
numpy==1.24.1
pandas>=1.3.0
opencv-python>=4.5.0
pillow>=8.0.0
scipy>=1.7.0
easyocr>=1.6.0
torch>=1.10.0
torchvision>=0.11.0
transformers==4.37.0

fastapi==0.115.5
opencv-python
tqdm==4.65.2
Pillow==9.3.0
uvicorn
decord==0.6.0
pandas
pyyaml==6.0.1
scikit-video
timm
einops
clip
opencv-python